% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{CK.gdist}
\alias{CK.gdist}
\title{Function to take two objects of class phylo and return their topological distance}
\usage{
CK.gdist(tr1, tr2)
}
\arguments{
\item{tr1}{...}

\item{tr2}{...}
}
\description{
Description of this function..
}

