% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{CK.gdistm}
\alias{CK.gdistm}
\title{Function to take an object of class multiPhylo and return topological distance matrix}
\usage{
CK.gdistm(trees)
}
\arguments{
\item{trees}{...}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

