% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{CK.metric}
\alias{CK.metric}
\title{Title of the function}
\usage{
CK.metric(x, lambda = 0, type = "number")
}
\arguments{
\item{x}{an object of the class \code{phylo} ...}

\item{lambda}{...}

\item{type}{...}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

