% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{CK.wdist}
\alias{CK.wdist}
\title{Function to take two objects of class phylo and return their weighted distance}
\usage{
CK.wdist(tr1, tr2, p = 1)
}
\arguments{
\item{tr1}{...}

\item{tr2}{...}

\item{p}{...}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

