% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{CK.wdistm}
\alias{CK.wdistm}
\title{Function to take an object of class multiPhylo and return weighted distance matrix}
\usage{
CK.wdistm(trees, p = 1)
}
\arguments{
\item{trees}{...}

\item{p}{...}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

