% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{CPP_update_combinations}
\alias{CPP_update_combinations}
\title{CPP leaf combination distance function}
\usage{
CPP_update_combinations(length_root_distances, topological_root_distances,
  left_partition, right_partition, index_offsets, distance_to_root,
  edges_to_root)
}
\arguments{
\item{length_root_distances}{The vector of length distances to the root (for each combination of leaves).}

\item{topological_root_distances}{The vector of edge distances to the root (for each combination of leaves).}

\item{left_partition}{A vector with the leaves on the left subtree of the current node.}

\item{right_partition}{A vector with the leaves on the right subtree of the current node.}

\item{index_offsets}{Auxiliary vector used to compute the vector index for a given combination.}

\item{distance_to_root}{The branch length distance (branch length to the root).}

\item{edges_to_root}{The edge distance (number of edges to the root).}
}
\description{
CPP implementation of the function that updates the vector of distances for a set of leaf combinations (the ones
induced by a specific node of the tree). We modify the vectors within the function (to avoid copying them).
}
\author{
Jacob Almagro-Garcia \email{nativecoder@gmail.com}
}

