% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{edge.pendant}
\alias{edge.pendant}
\title{Function to create a vector of the pendant edges of the tree}
\usage{
edge.pendant(tree, k)
}
\arguments{
\item{tree}{...}

\item{k}{...}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

