% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/exploratree.R
\name{exploratree}
\alias{exploratree}
\title{Phylogenetic tree exploration}
\usage{
exploratree(x, explorer = distTips, nf = NULL, ...)
}
\arguments{
\item{x}{}

\item{explorer}{}

\item{...}{}
}
\description{
This functions are under development. Please do not use them without contacting the author first.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}, Michelle Kendall \email{michelle.louise.kendall@gmail.com}, Caroline Colijn \email{c.colijn@imperial.ac.uk}
}

