% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/servers.R
\name{exploratreeServer}
\alias{exploratreeServer}
\title{Web-based tree explorer}
\usage{
exploratreeServer()
}
\description{
This function opens up an application in a web browser for an interactive exploration of the diversity in a set of trees. This function is under development. Please do not use without contacting the author.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

