% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clusters.R
\name{find.groves}
\alias{find.groves}
\title{Identify clusters of similar trees}
\usage{
find.groves(x, method = tree.vec, nf = NULL, clustering = "ward.D2",
  nclust = NULL, ...)
}
\arguments{
\item{x}{an object of the class multiPhylo}

\item{method}{a function outputting the summary of a tree (phylo object) in the form of a vector}

\item{nf}{the number of principal components to retain}

\item{clustering}{a character string indicating the clustering method to be used; see argument \code{method} in \code{?hclust} for more details.}

\item{...}{further arguments to be passed to \code{treescape}}
}
\value{
A list containing:
\itemize{
 \item groups: a factor defining groups of trees
 \item treescape: the output of treescape
}
}
\description{
This function uses hierarchical clustering on principal components output by \link{\code{treescape}} to identify groups of similar trees. Clustering relies on \link{\code{hclust}}, using Ward's method by default.
}
\examples{
if(require("adegenet")){
## load data
data(woodmiceTrees)

## use treescape
res <- find.groves(x, nf=5, nclust=6)
PCs <- res$treescape$pco$li
s.class(PCs, fac=res$groups, col=funky(6))
}
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

