% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{med.tree}
\alias{med.tree}
\title{Geometric median tree function}
\usage{
med.tree(trees, likes = rep(1, length(trees)), lambda = 0)
}
\arguments{
\item{trees}{an object of the class \code{multiPhylo}}

\item{likes}{a vector of weightings for the trees. Defaults to a vector of 1's so that all trees are equally weighted, but can be used to weight trees according to likelihood or other characteristics.}

\item{lambda}{a number in [0,1] which specifies the extent to which topology (default, with lambda=0)  or branch lengths (lambda=1) are emphasised. This argument is ignored if \code{type="function"}.}
}
\description{
Finding the geometric median of a set of trees according to the Kendall Colijn metric
}
\examples{
## generate 10 random trees, each with 6 tips
trees <- rmtree(10,6)
## Geometric median tree:
mymedian <- med.tree(trees)
mymedian$centre # the vector at the 'centre' of the trees; may not correspond to an actual tree
mymedian$median # the identifier(s) of the tree(s) closest to the central vector
mymedian$mindist # the distance of the median tree(s) from the central vector
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

