% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{my.mrca}
\alias{my.mrca}
\title{Function to make mrca matrix of a tree, where entry (i,j) gives the mrca of tips i and j}
\usage{
my.mrca(tree, k)
}
\arguments{
\item{tree}{...}

\item{k}{...
#'}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

