% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{pen.edge.treematch}
\alias{pen.edge.treematch}
\title{Title of the function}
\usage{
pen.edge.treematch(tree, labelmatch)
}
\arguments{
\item{tree}{...}

\item{labelmatch}{...}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

