% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{CK.metric}
\alias{CK.metric}
\title{Tree vector function}
\usage{
CK.metric(tr1, lambda = 0, type = "number")
}
\arguments{
\item{lambda}{...}

\item{type}{...}

\item{x}{an object of the class \code{phylo} ...}
}
\description{
Function which takes a phylo as input and outputs the vector.
The elements of the vector are numeric if a value for lambda is supplied and type="number",
otherwise they are functions of lambda
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

