% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{linear.mrca}
\alias{linear.mrca}
\title{Linear MRCA function}
\usage{
linear.mrca(tree, k = 0)
}
\arguments{
\item{tree}{object of class phylo}

\item{k}{number of tips in tree}
}
\description{
Function to make mrca matrix of a tree, where entry (i,j) gives the mrca of tips i and j
}
\examples{
## generate a phylo tree
x <- rtree(6)

## create matrix of MRCAs: entry (i,j) is the node number of the MRCA of tips i and j
linear.mrca(x,6)
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

