% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{med.tree}
\alias{med.tree}
\title{Geometric median tree function}
\usage{
med.tree(trees, lambda = 0, weights = rep(1, length(trees)),
  return_lambda_function = F, save_memory = F)
}
\arguments{
\item{trees}{An object of the class multiPhylo, containing the trees for which the median tree will be computed.}

\item{lambda}{a number in [0,1] which specifies the extent to which topology (default, with lambda=0)  or branch lengths (lambda=1) are emphasised. This argument is ignored if \code{return_lambda_function=TRUE}.}

\item{weights}{A vector of weights for the trees. Defaults to a vector of 1's so that all trees are equally weighted, but can be used to encode likelihood, posterior probabilities or other characteristics.}

\item{return_lambda_function}{If true, a function that can be invoked with different lambda values is returned.
This function returns the vector of metric values for the given lambda.}

\item{save_memory}{A flag that saves a lot of memory but increases the execution time (not compatible with return_lambda_function=TRUE).}
}
\value{
A list with the median metric vector, distances, indices of the tree(s) that are closest to the median tree and the value of this distance
 or a function that produces this list for a given value of lambda.
}
\description{
Finds the geometric median of a set of trees according to the Kendall Colijn metric.
}
\examples{
## generate 10 random trees, each with 6 tips
trees <- rmtree(10,6)
## Geometric median tree:
mymedian <- med.tree(trees)
mymedian$centre # the vector at the 'centre' of the trees; may not correspond to an actual tree
mymedian$median # the identifier(s) of the tree(s) closest to the central vector
mymedian$mindist # the distance of the median tree(s) from the central vector

\dontrun{
## Example with woodmice data:
data(woodmiceTrees)
woodmiceMed <- med.tree(woodmiceTrees)
## plot the (first) geometric median tree (there are seven topologically identical median trees):
plot(woodmiceTrees[[woodmiceMed$median[[1]]]])

## finding the geometric median tree from a single cluster:
woodmiceDists <- multi.dist(woodmiceTrees)
woodmiceMDS <- dudi.pco(as.dist(woodmiceDists), scannf=FALSE, nf=2)
## isolate the trees from the largest cluster
woodmiceCluster1 <- woodmiceTrees[intersect(
   intersect(which(woodmiceMDS$li[,1]>(-2)),which(woodmiceMDS$li[,1]<2)),
   intersect(which(woodmiceMDS$li[,2]>(-2.5)),which(woodmiceMDS$li[,1]<2.5))
   )]
## find the geometric median
geomMedWoodmice1 <- med.tree(woodmiceCluster1)
plot(woodmiceCluster1[[geomMedWoodmice1$median[[1]]]])
# this has the same topology as the overall median tree:
tree.dist(woodmiceTrees[[woodmiceMed$median[[1]]]],
  woodmiceCluster1[[geomMedWoodmice1$median[[1]]]])

## However, median trees from other clusters have different topologies, for example:
## isolate the trees from the second largest cluster:
woodmiceCluster2 <- woodmiceTrees[intersect(
 intersect(which(woodmiceMDS$li[,1]>(-1)),which(woodmiceMDS$li[,1]<8)),
 intersect(which(woodmiceMDS$li[,2]>1),which(woodmiceMDS$li[,1]<6))
)]
## find the geometric median
geomMedWoodmice2 <- med.tree(woodmiceCluster2)
plot(woodmiceCluster2[[geomMedWoodmice2$median[[1]]]])
## This is another representative topology, which is different from those we found above:
tree.dist(woodmiceCluster2[[geomMedWoodmice2$median[[1]]]],
  woodmiceCluster2[[geomMedWoodmice2$median[[1]]]])
}
}
\author{
Jacob Almagro-Garcia \email{nativecoder@gmail.com}

Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

