% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{multi.dist}
\alias{multi.dist}
\title{Metric function for \code{multiPhylo} input}
\usage{
multi.dist(trees, lambda = 0, type = "number")
}
\arguments{
\item{trees}{an object of the class \code{multiPhylo}}

\item{lambda}{a number in [0,1] which specifies the extent to which topology (default, with lambda=0)  or branch lengths (lambda=1) are emphasised. This argument is ignored if type="function"}

\item{type}{logical which takes the inputs "\code{number}" (default) or "\code{function}". When \code{type="number"}, the output is a matrix where each entry is numeric; when \code{type="function"} the output is a matrix where each entry is a function of lambda}
}
\description{
Comparison of a list of trees using the Kendall Colijn metric. Output is given as a pairwise distance matrix.
}
\examples{
## generate 10 random trees, each with 6 tips
trees <- rmtree(10,6)
## pairwise distance matrix when lambda=0
multi.dist(trees)
## pairwise distance matrix as a function of lambda:
m <- multi.dist(trees, type="function")
## evaluate at lambda=0. Equivalent to multi.dist(trees).
m(0)

## A method to visualise these distances with MDS:
require(ade4)
mMDS <- dudi.pco(as.dist(m(0)), scannf=FALSE,nf=2) # finds a projection of the points in 2 dimensions
mdf <- as.data.frame(cbind(mMDS$li[,1],mMDS$li[,2])) # put the coordinates of these points into a data frame
require(ggplot2)
mplot <- ggplot(mdf, aes(mMDS$li[,1],mMDS$li[,2])) # create plot
require(RColorBrewer)
mpalette <- brewer.pal(10,"Paired") # create colour palette
## plot:
mplot + geom_point(colour=mpalette,size=5) +
  xlab("") + ylab("") + theme_bw(base_family = "")
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

