% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{pen.edge.tree}
\alias{pen.edge.tree}
\title{Pendant edges}
\usage{
pen.edge.tree(tree, k)
}
\arguments{
\item{tree}{object of class phylo}

\item{k}{number of tips in tree}
}
\description{
Extract just the pendant edges from the vector tree$edge
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

