% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{pen.edge.treematch}
\alias{pen.edge.treematch}
\title{Pendant edges, matched}
\usage{
pen.edge.treematch(tree, labelmatch)
}
\arguments{
\item{tree}{an object of the class \code{phylo}}

\item{labelmatch}{a vector specifying the order of the tips in the output. This is used by other functions in the package to match the tip labels in the trees}
}
\description{
Extract the pendant edges from the vector \code{tree$edge}, in the order given by \code{labelmatch}.
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

