% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metrics.R
\name{pen.edge.treematch}
\alias{pen.edge.treematch}
\title{Pendant edges, matched}
\usage{
pen.edge.treematch(tree, labelmatch)
}
\arguments{
\item{tree}{...}

\item{labelmatch}{...}
}
\description{
Extract the pendant edges from the vector tree$edge, in the order given by labelmatch
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

