% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metrics.R
\name{tree.dist}
\alias{tree.dist}
\title{Title of the function}
\usage{
tree.dist(tr1, tr2, lambda = 0, type = "number")
}
\arguments{
\item{tr1}{...}

\item{tr2}{...}

\item{lambda}{...}

\item{type}{...}
}
\description{
Description of this function..
}
\author{
Michelle Kendall \email{michelle.louise.kendall@gmail.com}
}

