% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/servers.R
\name{treescapeServer}
\alias{treescapeServer}
\title{Web-based tree explorer}
\usage{
treescapeServer()
}
\description{
This function opens up an application in a web browser for an interactive exploration of the diversity in a set of trees.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}

